beginning of breaking block goal
the start of the breaking block goal, it is bugged and doesnt even do anything lmao but it's a start. Thanks Harby for the help :]

+++ src/main/java/mod/pilot/steampunked/entity/reconstucted/POI.java
@@ -28,7 +28,7 @@ public class POI {
         this.POIPos = block.getCenter();
     }
     public POI(ReconstructedBase parent, Mob target){
-        int priorityCheck = (int)Steampunked.GetMobPriorityFor(target);
+        int priorityCheck = Steampunked.GetMobPriorityFor(target);
         if (priorityCheck == 1){
             this.Priority = 3;
         }
